(wip) add copy curl button to test run page
