+++ src/resources/js/pages/sessions/test-steps/index.vue
@@ -73,12 +73,10 @@
                                     hide-header-close
                                     title="Request"
                                 >
-                                    <clipboard-copy-btn
-                                        :data="jsonToCurl(testStep.request)"
-                                        :show-copy-icon="false"
-                                        :copy-button-text="'Copy cURL'"
-                                        style="position: inherit; top: -.5rem; left: 35rem;"
-                                    ></clipboard-copy-btn>
+                                    <clipboard-json-to-curl
+                                        :request="testStep.request"
+                                    >
+                                    </clipboard-json-to-curl>
                                     <div class="border">
                                         <div
                                             class="d-flex"
@@ -297,18 +295,5 @@ export default {
             required: true,
         },
     },
-    methods: {
-        jsonToCurl: function (request) {
-            let result = 'curl --location --request ';
-
-            result = result.concat(request['method']," 'Session_URL",request['uri'],"' \\  ")
-            for (let h in request['headers']) {
-                result += "--header '"+h+": "+ request['headers'][h][0]+"' \\  "
-            }
-            result = result.concat("--data-raw '",JSON.stringify(request['body']), "'")
-
-            return result;
-        }
-    }
 };
 </script>
